Add ability to use "custom" workspaces.
This is useful if you have a pre-build script that builds things into
another subfolder. In that scenario, you might still want `workspaces`
in `package.json` in order for `yarn install` to work correctly but the
thing you actually publish to npm comes from a different subfolder.

Specifically, this allows glimmer-vm to leverage this publishing script
without a massive refactor of its build pipeline.