Prompt if publishing scoped package fails with private error.
Publishing a scoped package defaults to `--access private` on the NPM
registry for new packages (unless the package has setup a
`publishConfig` property to indicate public access).

This commit detects that specific failure and prompts the user to see if
they really meant to publish it as a public package. There is no default
value for the confirmation prompt (meaning you **must** answer either
"Yes" or "No") in order to avoid an accidental private information leak.