Ensure prompts have access to required information.
When using `this.step` `release-it` internally uses `this.global.isCI`
to switch between `this.spinner.show` (which is mocked out to
automatically invoke the task property on `options` in its test helpers)
and `this.prompt.show`.

Unfortunately, there is no _good_ way to test these prompts (that I can
see). :think: