Create unified set of plugin operations.
This makes asserting on specific event ordering much easier. Previously,
the only thing we asserted on was that the commands all ran in the right
order, and the logs were all in the right order, and the prompts were
all in the right order. Now we will be able to assert that the commands,
logs, and prompts are in the right order in relationship to _each
other_.