Fix `pnpm` publishing (and `publishCommand` configuration) (#127)
The prior implementation was incorrect. I had thought that the 
underlying `Plugin.prototype.exec` internally used lodash 
templating, but it doesn't. 

This fixes publishing in general (with either `npm` or `pnpm`),
as well as changing the approach to how to customize the publish
command. Now you have to write a script that leverages
environment variables (instead of templating / interpolation).